+++ hutool-core/src/main/java/cn/hutool/core/comparator/VersionComparator.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.comparator;
 
 import cn.hutool.core.convert.Convert;
+import cn.hutool.core.lang.Version;
 import cn.hutool.core.util.*;
 
 import java.io.Serializable;
@@ -13,15 +14,13 @@ import java.util.regex.Pattern;
  * 比较两个版本的大小<br>
  * 排序时版本从小到大排序，即比较时小版本在前，大版本在后<br>
  * 支持如：1.3.20.8，6.82.20160101，8.5a/8.5c等版本形式<br>
- * 参考：https://www.cnblogs.com/shihaiming/p/6286575.html
+ * 参考：java.lang.module.ModuleDescriptor.Version
  *
  * @author Looly
  * @since 4.0.2
  */
 public class VersionComparator implements Comparator<String>, Serializable {
 	private static final long serialVersionUID = 8083701245147495562L;
-
-	private static final Pattern PATTERN_PRE_NUMBERS= Pattern.compile("^\\d+");
 
 	/** 单例 */
 	public static final VersionComparator INSTANCE = new VersionComparator();
@@ -64,39 +63,6 @@ public class VersionComparator implements Comparator<String>, Serializable {
 			return 1;
 		}
 
-		final List<String> v1s = StrUtil.split(version1, CharUtil.DOT);
-		final List<String> v2s = StrUtil.split(version2, CharUtil.DOT);
-
-		int diff = 0;
-		int minLength = Math.min(v1s.size(), v2s.size());// 取最小长度值
-		String v1;
-		String v2;
-		for (int i = 0; i < minLength; i++) {
-			v1 = v1s.get(i);
-			v2 = v2s.get(i);
-			// 先比较长度
-			diff = v1.length() - v2.length();
-			if (0 == diff) {
-				diff = v1.compareTo(v2);
-			}else {
-				// 不同长度，且含有字母
-				if(!NumberUtil.isNumber(v1) || !NumberUtil.isNumber(v2)){
-					//不同长度的先比较前面的数字；前面数字不相等时，按数字大小比较；数字相等的时候，继续按长度比较，类似于 103 > 102a
-					final int v1Num = Convert.toInt(ReUtil.get(PATTERN_PRE_NUMBERS, v1, 0), 0);
-					final int v2Num = Convert.toInt(ReUtil.get(PATTERN_PRE_NUMBERS, v2, 0), 0);
-					final int diff1 = v1Num - v2Num;
-					if (diff1 != 0) {
-						diff = diff1;
-					}
-				}
-			}
-			if(diff != 0) {
-				//已有结果，结束
-				break;
-			}
-		}
-
-		// 如果已经分出大小，则直接返回，如果未分出大小，则再比较位数，有子版本的为大；
-		return (diff != 0) ? diff : v1s.size() - v2s.size();
+		return CompareUtil.compare(Version.of(version1), Version.of(version2));
 	}
 }